+++ src/main/kotlin/crimera/patches/twitter/misc/settings/SettingsResourcePatch.kt
@@ -7,8 +7,8 @@ import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
 import app.revanced.util.ResourceGroup
 import app.revanced.util.copyResources
-import org.w3c.dom.Element
 import app.revanced.util.copyXmlNode
+import org.w3c.dom.Element
 
 @Patch(
     compatiblePackages = [CompatiblePackage("com.twitter.android")],
@@ -23,7 +23,7 @@ object SettingsResourcePatch: ResourcePatch() {
 
             val prefMod = editor.file.createElement("Preference")
             prefMod.setAttribute("android:icon", "@drawable/ic_vector_settings_stroke")
-            prefMod.setAttribute("android:title", "@string/piko_settings_title")
+            prefMod.setAttribute("android:title", "@string/piko_title_settings")
             prefMod.setAttribute("android:key", "pref_mod")
             prefMod.setAttribute("android:order", "110")
 
@@ -34,7 +34,7 @@ object SettingsResourcePatch: ResourcePatch() {
             val applicationNode = it.file.getElementsByTagName("application").item(0)
 
             val modActivity = it.file.createElement("activity").apply {
-                setAttribute("android:label", "@strings/piko_settings_title")
+                setAttribute("android:label", "@strings/piko_title_settings")
                 setAttribute("android:name", "app.revanced.integrations.twitter.settings.SettingsActivity")
                 setAttribute("android:excludeFromRecents", "true")
             }
@@ -48,16 +48,18 @@ object SettingsResourcePatch: ResourcePatch() {
         /**
          * create directory for the untranslated language resources
          */
-        context["res/values-v21"].mkdirs()
-        arrayOf(
-            ResourceGroup(
-                "values-v21",
-                "strings.xml"
-            )
-        ).forEach { resourceGroup ->
-            context.copyResources("twitter/settings", resourceGroup)
+        val languages = arrayOf(
+            "ru",
+            "v21"
+        ).map { "values-$it" }
+
+        languages.forEach {
+            if (context["res/$it"].exists()) {
+                context.copyXmlNode("twitter/settings", "$it/strings.xml", "resources")
+            } else {
+                context["res/$it"].mkdirs()
+                context.copyResources("twitter/settings", ResourceGroup(it, "strings.xml"))
+            }
         }
-
-
     }
 }